separated useInput
